s specifier for printf done

+++ ft_printf/sources/ft_accuracy.c
@@ -6,17 +6,17 @@
 /*   By: marrow <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/08 23:28:56 by marrow            #+#    #+#             */
-/*   Updated: 2020/02/12 14:24:27 by marrow           ###   ########.fr       */
+/*   Updated: 2020/02/18 17:49:56 by marrow           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	    	ft_accuracy(t_spec *specifier, size_t length)
+void	    	ft_accuracy(t_spec *specifier, int length)
 {
-	if (length >= specifier->accuracy)
+	if (length >= (int)specifier->accuracy)
 		return ;
-	length = specifier->accuracy - length;
+	length = (int)specifier->accuracy - length;
 	while (length-- > 0)
 		ft_putchar('0');
 }
